Backend/requirements.txt
# FastAPI & Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# MongoDB
motor>=3.3.0
pymongo>=4.6.0

# HTTP Client
httpx>=0.25.0

# Scheduling
apscheduler>=3.10.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Machine Learning / Statistics
scikit-learn>=1.3.0
statsmodels>=0.14.0
joblib>=1.3.0  # For model persistence

# GTFS-Realtime (MTA)
gtfs-realtime-bindings>=1.0.0
protobuf>=4.24.0

# Hugging Face
huggingface-hub>=0.19.0
transformers>=4.35.0  # Optional, for local models

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dateutil>=2.8.2

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
